fix(Tests): renamed util file
